+++ lib/dawn/knowledge_base.rb
@@ -229,6 +229,7 @@ require "dawn/kb/osvdb_108570"
 require "dawn/kb/osvdb_108530"
 require "dawn/kb/osvdb_108563"
 require "dawn/kb/osvdb_118579"
+require "dawn/kb/osvdb_118830"
 require "dawn/kb/osvdb_118954"
 require "dawn/kb/osvdb_119878"
 require "dawn/kb/osvdb_119927"
@@ -483,6 +484,7 @@ module Dawn
           Dawn::Kb::OSVDB_108530.new,
           Dawn::Kb::OSVDB_108563.new,
           Dawn::Kb::OSVDB_118579.new,
+          Dawn::Kb::OSVDB_118830.new,
           Dawn::Kb::OSVDB_118954.new,
           Dawn::Kb::OSVDB_119878.new,
           Dawn::Kb::OSVDB_119927.new,